Replace pytorch geometric with AtomicData (#1156)
* pull over Xiangs atomic data

* rename uma

* working uma speed benchmark

* progress

* debug

* fix test issues, need tf32 context

* lint and fixes

* stash pop

* failing on tensor property

* remove tensor property for now

* support dummy binary dataset injection into atomic_data

* patch through dummy dataset for element ref test

* fix tests

* convert to using partial

* fix tests

* lint and format

* migrate collater

* clean up

* pbc for v2

* remove unused atoms_to_graphs

* remove full atomstographs

* lint and fix tests

* fixes

* rever to original values

* replace Batch/Data

* lint

* remove dtype of tags, since atomicdata already validates this

* fix gpu tests

* skip finetune test for now

* remove lmdb_dataset and debug lines

* remove debug lines

* cleanup

* fix test

* skip another test

* move away from pickle/pyg

* fix fake dataset path

* fix test

+++ tests/core/units/mlip_unit/create_fake_dataset.py
@@ -110,7 +110,7 @@ def generate_structures(fake_dataset_config: FakeDatasetConfig):
         atom_positions = np.random.uniform(0, sys_size, (n_atoms, 3))
         if fake_dataset_config.pbc:
             pbc = True
-            cell = np.eye(3) * sys_size
+            cell = np.eye(3, dtype=np.float32) * sys_size
         else:
             pbc = False
             cell = None
@@ -122,6 +122,7 @@ def generate_structures(fake_dataset_config: FakeDatasetConfig):
         forces = calculate_forces_repulsive_force_field(atoms)
         energy = compute_energy(atoms)
         systems.append({"atoms": atoms, "forces": forces, "energy": energy})
+
     forces = torch.vstack([system["forces"] for system in systems])
     energies = torch.vstack([system["energy"] for system in systems])
 
@@ -153,6 +154,7 @@ def generate_structures(fake_dataset_config: FakeDatasetConfig):
         atoms.info["spin"] = np.random.randint(0, 2)
 
         structures.append(atoms)
+
     return structures
